// apps/web/src/lib/api/generated/main/model/event-coordination-response-dto-heroes-item.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { EventCoordinationResponseDtoHeroesItemActiveGapsItem } from './event-coordination-response-dto-heroes-item-active-gaps-item.js';
import type { EventCoordinationResponseDtoHeroesItemCoverage } from './event-coordination-response-dto-heroes-item-coverage.js';
import type { EventCoordinationResponseDtoHeroesItemPriority } from './event-coordination-response-dto-heroes-item-priority.js';
import type { EventCoordinationResponseDtoHeroesItemRecommendedAction } from './event-coordination-response-dto-heroes-item-recommended-action.js';
import type { EventCoordinationResponseDtoHeroesItemTimer } from './event-coordination-response-dto-heroes-item-timer.js';

export type EventCoordinationResponseDtoHeroesItem = {
  heroId: string;
  /** @nullable */
  npcId: number | null;
  npcName: string;
  /** @nullable */
  npcIcon: string | null;
  /** @nullable */
  npcLvl: number | null;
  /** @nullable */
  timer: EventCoordinationResponseDtoHeroesItemTimer;
  coverage: EventCoordinationResponseDtoHeroesItemCoverage;
  activeGaps: EventCoordinationResponseDtoHeroesItemActiveGapsItem[];
  priority: EventCoordinationResponseDtoHeroesItemPriority;
  recommendedAction: EventCoordinationResponseDtoHeroesItemRecommendedAction;
};